# WIP: implement #4916, keyword args as dictionary