--- a/test/parse.jl
+++ b/test/parse.jl
@@ -989,7 +989,7 @@ let f = function (x; kw...)
     g = function (x; a = 2)
             return (x, a)
         end
-    @test f(1) == (1, Any[])
+    @test f(1) == (1, Base.KWDict())
     @test g(1) == (1, 2)
 end