This seems like a good change to me, and while some details are up in the air the important thing is the API change (namely, rest keywords being dict-like and immutable). Haven't run all tests yet, but this went pretty smoothly so far.

I think immutability really makes sense here since there tend to be few keyword args, and function arguments tend to be used in an immutable way. Positional varargs are of course already immutable. This already eliminates some allocations (unnecessary empty `Vector{Any}`s), and at least for small argument counts I expect this to speed up the hard-to-optimize uses of keyword args.

I'm not sure what type to use here but `ImmutableDict{Symbol,Any}` seems to work well for now.

Previously, rest keywords were sloppy about duplicates:

```
julia> ks(;x...) = x;

julia> ks(;[:a=>1]..., [:a=>2]...)
2-element Array{Any,1}:
 (:a, 1)
 (:a, 2)
```

With this change, we use the constructed dict to de-duplicate the keyword names (keeping only the last one):

```
julia> ks(;[:a=>1]..., [:a=>2]...)
Base.ImmutableDict{Symbol,Any} with 1 entry:
  :a => 2
```

However, this is potentially costly since `ImmutableDict` lookup is O(n).

At the same time, rest keywords need to preserve the order they were passed in, and `ImmutableDict` is perfect for that since it happens to be ordered. I know I've written code that assumes keyword arg order is preserved, but I guess we can consider dropping that guarantee.