@JeffBezanson 2017-05-16
Highly-biased benchmark:

Before:

```
f(;kw...) = g(;kw...)
g(;kw...) = kw
c = g(a=1,b=2)

julia> @time for i = 1:10000; f(a=1,b=2); end
  0.034662 seconds (140.00 k allocations: 7.019 MiB)

julia> @time for i = 1:10000; f(;c...); end
  0.056522 seconds (180.00 k allocations: 8.240 MiB)
```

after:

```
julia> @time for i = 1:10000; f(a=1,b=2); end
  0.003358 seconds (30.00 k allocations: 1.678 MiB)

julia> @time for i = 1:10000; f(;c...); end
  0.000678 seconds
```