@alyst 2017-05-16
:+1: Does it also improve type inference (e.g. #21811, #21794)?