@JeffBezanson 2017-05-16
No this has no effect on type inference.