@JeffBezanson 2017-05-16
You're still allowed to pass duplicate keyword arguments, but they get de-duplicated. However the implementation of `ImmutableDict` allows multiple mappings for the same key, so with the current state of this PR it's possible to just remove the de-duplication if we want. _However_ that doesn't fit so well with the `args[name]` dictionary interface, and might be harder to support with other implementations of the keyword args container.