@andyferris 2017-05-17
Cool!

I've been playing with the idea of using a strongly-typed immutable dictionary for the keyword arguments, with a proof-of-principle demonstration using macros at the definition and call sites (this is still early WIP but I'm confident it would work) with the idea that lowering could implement the macro behavior automatically...

So I *think* we could have all the things (dictionary structure for the varargs (rest?) keywords, *plus* type stability) without too much fuss - the main difference to this being using a strongly-typed dict instead of `ImmutableDict`.