@JeffBezanson 2017-05-17
Yes, I think it's likely we will eventually use whatever form of built-in NamedTuple type we end up with for this.

Here it gets interesting, since from my experience with NamedTuples (and also true of namedtuples in python), the elements should just be the values, and not key=>value pairs, since the names are part of the structure (or type), and not part of the data. That raises the question of what the interface for keyword argument providers should be. Currently we expect them to be (1) iterable, and (2) generate elements that can be iterated to yield 2 values (a name Symbol and a value). That works, but it seems ad-hoc and inelegant to me, and doesn't fit well with more strongly-typed containers where the type of the value can depend on the name. Anybody have any ideas?