@ajkeller34 2017-05-18
I'm not sure if this is a good idea, but since this PR suggests using an `ImmutableDict` as a keyword argument provider, here are the bones of a type-stable `ImmutableDict`-like object, along with a simple example showing that it remains type-stable when a duplicate key is entered with an associated value of a different type than before. I something like this could work if passed to a `kwsorter` (provided iteration, etc. were implemented). 

https://gist.github.com/ajkeller34/9987cfcc53660cfbb5349965fc6a2a52

Then again, the type signatures become pretty ungainly, so the elegance criterion may not be satisfied.