@JeffBezanson 2017-05-19
Thanks Andrew, that's pretty clever. I think ultimately we will have to use a standard NamedTuple type (ala #16580) for this.