@quinnj 2017-05-23
So is the idea to have a fully baked NamedTuple first and then implement this using that? Or make this change 1st w/ NamedTuple at some future date?